refactor(governance): the two static guards walk the tree through one shared module
Review finding: the new guard copied `APP_DIR`/`SRC_DIR`/`EE_DIR`, `EXTENSIONS`
and the source-file walker byte-for-byte out of the boundary guard sitting in
the same directory. CLAUDE.md's reuse rule is explicit, and the precedent for
sharing between test files in one directory is an underscore-prefixed module
(`src/server/nlpgo/__tests__/_nlpgoBinaryStamp.ts`).

Both guards fail in the silent direction — a walker that returns fewer files
than it should reports the same clean result as a codebase with nothing to
find — so a correction to what "every source file" means has to land in both at
once, which one copy cannot promise.

The on-demand guard's self-check (the comparator's key must still be found in
the two files that carry it) is what proves the shared walker still reads.
